requirements.txt
asgiref>=3.8.1
certifi>=2024.7.4
distlib>=0.3.8
Django>=5.0.7
django-debug-toolbar>=4.4.6
django-filter>=24.2
djangorestframework>=3.15.2
drf-nested-routers>=0.94.1
filelock>=3.15.4
mysqlclient>=2.2.4
pipenv>=2024.0.1
platformdirs>=4.2.2
Pygments>=2.18.0
setuptools>=71.1.0
sqlparse>=0.5.1
tzdata>=2024.1
virtualenv>=20.26.3
